Lint.

+++ chia/rpc/data_layer_rpc_api.py
@@ -243,9 +243,12 @@ class DataLayerRpcApi:
         if self.service is None:
             raise Exception("Data layer not created")
         store_id = request.get("id")
+        if store_id is None:
+            raise Exception("missing store id in request")
         store_id_bytes = bytes32.from_hexstr(store_id)
         urls = request["urls"]
         await self.service.remove_subscriptions(store_id=store_id_bytes, urls=urls)
+        return {}
 
     async def add_missing_files(self, request: Dict[str, Any]) -> Dict[str, Any]:
         """
